app/requirements.txt
gradio
numpy
pathlib
passlib
pydicom
Pillow
git+https://github.com/qubvel/segmentation_models.pytorch
ttach
psutil
scikit-learn
scikit-image
scikit-build
matplotlib
opencv-python
pandas
bcrypt
awscli
boto3